# Ease of adding new routes to the Vertx router
We would like to add some custom routes to the Vertx router, but its prototype bean scope results in a new instance per injection point, so there is no way
to modify the one injected in PrebidVerticle.
It would be handy if WebConfiguration could receive an optional Consumer<Router> that would allow us to add these new routes.